Another half piece of the puzzle, receive the PublishedEvent in the ESB internals
