+++ akka-persistence-typed/src/main/scala/akka/persistence/typed/internal/BehaviorSetup.scala
@@ -14,7 +14,6 @@ import akka.actor.typed.scaladsl.ActorContext
 import akka.actor.typed.ActorRef
 import akka.annotation.InternalApi
 import akka.persistence._
-import akka.persistence.typed.PublishedEvent
 import akka.persistence.typed.scaladsl.EventSourcedBehavior.ActiveActive
 import akka.persistence.typed.{ EventAdapter, PersistenceId, SnapshotAdapter }
 import akka.persistence.typed.scaladsl.{ EventSourcedBehavior, RetentionCriteria }
@@ -63,6 +62,8 @@ private[akka] final class BehaviorSetup[C, E, S](
 
   val journal: ClassicActorRef = persistence.journalFor(settings.journalPluginId)
   val snapshotStore: ClassicActorRef = persistence.snapshotStoreFor(settings.snapshotPluginId)
+
+  val replicaId: Option[String] = activeActive.map(_.replicaId)
 
   def selfClassic: ClassicActorRef = context.self.toClassic
